Relocate orchestrator e2e helpers into support utils.
This moves shared helper modules out of specs into support/utils and updates spec imports to follow workspace structure conventions while preserving backwards-compatible re-exports.
